change book price to int; rework getBooksSortedByAvailability, fix getBooksSortedByName.

+++ src/main/java/com/andersentask/bookshop/entities/Book.java
@@ -27,5 +27,5 @@ public class Book {
     private BookStatus status;
 
     @Column
-    private int price;
+    private Double price;
 }
